move zap flags internally
